Add a Posty 1.x importer
This importer copies all of the media, templates, and pages varbatim,
and then imports the pages making blurbs out of the first paragraph of
each post, as well as updating the metadata as needed.

Now that we have this importer, the `posty import posty1 PATH` command
has been implemented.

I also had to implement a decent amount `Site` to get this to work. \o/