Add new page, new post functionality
Running `posty new page` or `posty new post` will copy the object from
the skeleton directory into the destination appropriate with that object
type, and will name the file with the slug of the name you give it.

Simply copies the skeleton file, so the name inside the file doesn't
match yet.